Moved contracts to application
