+++ application/bootstrap/contracts/contracts.go
@@ -20,22 +20,22 @@ import (
 	"fmt"
 
 	"github.com/insolar/insolar/application"
+	"github.com/insolar/insolar/application/builtin/contract/account"
+	"github.com/insolar/insolar/application/builtin/contract/costcenter"
+	"github.com/insolar/insolar/application/builtin/contract/deposit"
+	"github.com/insolar/insolar/application/builtin/contract/member"
+	"github.com/insolar/insolar/application/builtin/contract/migrationadmin"
+	"github.com/insolar/insolar/application/builtin/contract/migrationdaemon"
+	"github.com/insolar/insolar/application/builtin/contract/migrationshard"
+	"github.com/insolar/insolar/application/builtin/contract/nodedomain"
+	"github.com/insolar/insolar/application/builtin/contract/pkshard"
+	"github.com/insolar/insolar/application/builtin/contract/rootdomain"
+	"github.com/insolar/insolar/application/builtin/contract/wallet"
+	maProxy "github.com/insolar/insolar/application/builtin/proxy/migrationshard"
+	pkProxy "github.com/insolar/insolar/application/builtin/proxy/pkshard"
 	"github.com/insolar/insolar/application/genesisrefs"
 	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/account"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/costcenter"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/deposit"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/member"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/migrationadmin"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/migrationdaemon"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/migrationshard"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/nodedomain"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/pkshard"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/rootdomain"
-	"github.com/insolar/insolar/logicrunner/builtin/contract/wallet"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	maProxy "github.com/insolar/insolar/logicrunner/builtin/proxy/migrationshard"
-	pkProxy "github.com/insolar/insolar/logicrunner/builtin/proxy/pkshard"
 )
 
 func RootDomain(pkShardCount int) application.GenesisContractState {
