+++ application/builtin/contract/account/account.go
@@ -20,12 +20,12 @@ import (
 	"fmt"
 	"math/big"
 
+	"github.com/insolar/insolar/application/builtin/proxy/costcenter"
+	"github.com/insolar/insolar/application/builtin/proxy/deposit"
+	"github.com/insolar/insolar/application/builtin/proxy/member"
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation/safemath"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/costcenter"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/deposit"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/member"
 )
 
 type Account struct {
