+++ application/builtin/contract/migrationdaemon/migrationdaemon.go
@@ -20,12 +20,12 @@ import (
 	"fmt"
 	"math/big"
 
+	"github.com/insolar/insolar/application/builtin/proxy/deposit"
+	"github.com/insolar/insolar/application/builtin/proxy/member"
+	"github.com/insolar/insolar/application/builtin/proxy/migrationadmin"
+	"github.com/insolar/insolar/application/builtin/proxy/wallet"
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/deposit"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/member"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/migrationadmin"
-	"github.com/insolar/insolar/logicrunner/builtin/proxy/wallet"
 )
 
 const CONVERSION = "10"
