+++ application/builtin/initialization.go
@@ -22,19 +22,19 @@ package builtin
 import (
 	"github.com/pkg/errors"
 
-	account "github.com/insolar/insolar/logicrunner/builtin/contract/account"
-	costcenter "github.com/insolar/insolar/logicrunner/builtin/contract/costcenter"
-	deposit "github.com/insolar/insolar/logicrunner/builtin/contract/deposit"
-	helloworld "github.com/insolar/insolar/logicrunner/builtin/contract/helloworld"
-	member "github.com/insolar/insolar/logicrunner/builtin/contract/member"
-	migrationadmin "github.com/insolar/insolar/logicrunner/builtin/contract/migrationadmin"
-	migrationdaemon "github.com/insolar/insolar/logicrunner/builtin/contract/migrationdaemon"
-	migrationshard "github.com/insolar/insolar/logicrunner/builtin/contract/migrationshard"
-	nodedomain "github.com/insolar/insolar/logicrunner/builtin/contract/nodedomain"
-	noderecord "github.com/insolar/insolar/logicrunner/builtin/contract/noderecord"
-	pkshard "github.com/insolar/insolar/logicrunner/builtin/contract/pkshard"
-	rootdomain "github.com/insolar/insolar/logicrunner/builtin/contract/rootdomain"
-	wallet "github.com/insolar/insolar/logicrunner/builtin/contract/wallet"
+	account "github.com/insolar/insolar/application/builtin/contract/account"
+	costcenter "github.com/insolar/insolar/application/builtin/contract/costcenter"
+	deposit "github.com/insolar/insolar/application/builtin/contract/deposit"
+	helloworld "github.com/insolar/insolar/application/builtin/contract/helloworld"
+	member "github.com/insolar/insolar/application/builtin/contract/member"
+	migrationadmin "github.com/insolar/insolar/application/builtin/contract/migrationadmin"
+	migrationdaemon "github.com/insolar/insolar/application/builtin/contract/migrationdaemon"
+	migrationshard "github.com/insolar/insolar/application/builtin/contract/migrationshard"
+	nodedomain "github.com/insolar/insolar/application/builtin/contract/nodedomain"
+	noderecord "github.com/insolar/insolar/application/builtin/contract/noderecord"
+	pkshard "github.com/insolar/insolar/application/builtin/contract/pkshard"
+	rootdomain "github.com/insolar/insolar/application/builtin/contract/rootdomain"
+	wallet "github.com/insolar/insolar/application/builtin/contract/wallet"
 
 	XXX_insolar "github.com/insolar/insolar/insolar"
 	XXX_artifacts "github.com/insolar/insolar/logicrunner/artifacts"
